pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docmind"
version = "0.1.0"
description = "A Python-based documentation generator with offline LLM support"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "DocMind Team" }
]
keywords = ["documentation", "llm", "code-analysis", "markdown"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
    "openai>=1.0.0",
    "rich>=13.0.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
cuda = [
    "faiss-gpu>=1.7.0",
]

[project.scripts]
docmind = "docmind.cli:app"

[project.urls]
Homepage = "https://github.com/docmind/docmind"
Documentation = "https://github.com/docmind/docmind#readme"
Repository = "https://github.com/docmind/docmind"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"